+++ resources/classes/SampleObject.class.php
@@ -1,9 +1,8 @@
 <?php
-class SampleObject extends Database\DBObject {
-	function getKeyField() {return 'primary_key';}
-	function getTableName() {return 'table_name';}
+	class SampleObject extends \Database\DBObject {
+		function getKeyField() {return 'primary_key';}
+		function getTableName() {return 'table_name';}
 
-	public function __construct() {
+		public function __construct() {
 
-	}
-}
+		}