+++ resources/themes/default/footer.tpl.php
@@ -1,4 +1,4 @@
-	<?php foreach($app->document->getJavascript() as $script): ?>
+	<?php foreach(DOM\Document::getJavascript() as $script): ?>
 		<script type="text/javascript" src="<?php print $script; ?>"></script>
 	<?php endforeach; ?>
 	</body>