+++ resources/themes/default/functions.php
@@ -1,5 +1,7 @@
 <?php
-	$app->document->addJavascript( Helper::url("/public/stylesheets/jquery-1.11.1.min.js") );
-	
-	$app->document->addStylesheet( Helper::url("/public/stylesheets/screen.css") );
+	DOM\Document::addJavascript( Helper::url("/public/stylesheets/jquery-1.11.1.min.js") );
+	DOM\Document::addJavascript( Helper::url("/test/test.hs"), "head");
+
+
+	DOM\Document::addStylesheet( Helper::url("/public/stylesheets/screen.css") );
 ?>