+++ resources/themes/default/header.tpl.php
@@ -2,7 +2,7 @@
 <html>
 	<head>
 		<title><?php print $app->getTitle(); ?></title>
-		<?php foreach($app->document->getStylesheets() as $style): ?> 
+		<?php foreach(DOM\Document::getStylesheets() as $style): ?> 
 			<link rel="stylesheet" type="text/css" href="<?php print $style; ?>" media="all" />
 		<?php endforeach; ?>
 	</head>